# BE/api-spec.yml
openapi: "3.0.2"
info:
  title: Anti Retake API
  version: '1.2'
servers:
  - url: https://antiretake.xyz/api
paths:
...
  /users/folders/personal:
    get:
      summary: Get personal folder list of Current User.
      tags:
        - Profile
      operationId: API-071
      parameters:
        - in: query
          name: searchText
          required: true
          schema:
            type: string
        - in: query
          name: limit
          required: true
          schema:
            type: number
        - in: query
          name: classId
          schema:
            type: string
      responses:
        "200":
          description: Folder list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderList"
...
  /sets/{id}/learn:
    post:
      summary: Set learned study set by of current user.
      tags:
        - Profile
      operationId: API-074
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Successfully.
...
  /search/classes:
    get:
      summary: Get class list of search by keyword, filter and sorted.
      tags:
        - Search
      operationId: API-018
      parameters:
        - in: query
          name: searchText
          required: true
          schema:
            type: string
        - in: query
          name: schools
          schema:
            type: array
            items:
              type: string
        - in: query
          name: sortedBy
          schema:
            type: string
      responses:
        "200":
          description: Class list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassList"
  /search/users:
    get:
      summary: Get user list of search by keyword.
      tags:
        - Search
      operationId: API-019
      parameters:
        - in: query
          name: searchText
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Class list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
  /search/questions:
    get:
      summary: Get question list of search by keyword, filter and sorted.
      tags:
        - Search
      operationId: API-020
      parameters:
        - in: query
          name: searchText
          required: true
          schema:
            type: string
        - in: query
          name: textbooks
          schema:
            type: array
            items:
              type: string
        - in: query
          name: subjects
          schema:
            type: array
            items:
              type: string
        - in: query
          name: schools
          schema:
            type: array
            items:
              type: string
        - in: query
          name: sortedBy
          schema:
            type: string
      responses:
        "200":
          description: Question list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionList"
...
components:
  schemas:
...
    FolderCardInfo:
      type: object
      properties:
        id:
          type: string
        numOfSets:
          type: number
        name:
          type: string
        createdTime:
          type: string
    FolderList:
      type: array
      items:
        $ref: "#/components/schemas/FolderCardInfo"
    ClassCardInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        school:
          type: string
        createdTime:
          type: string
        numOfSets:
          type: number
        numOfMembers:
          type: number
    ClassList:
      type: array
      items:
        $ref: "#/components/schemas/ClassCardInfo"
...
    QuestionCardInfo:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        createdTime:
          type: string
        ownerAvaUrl:
          type: string
        ownerName:
          type: string
        content:
          type: string
        numOfUpvotes:
          type: number
        numOfDownvotes:
          type: number
        numOfComments:
          type: number
        textbooks:
          type: array
          items:
            type: string
        subjects:
          type: array
          items:
            type: string
        schools:
          type: array
          items:
            type: string
    QuestionList:
      type: array
      items:
        $ref: "#/components/schemas/QuestionCardInfo"
    UserCardInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avaUrl:
          type: string
        numOfClasses:
          type: number
        numOfSets:
          type: number
    UserList:
      type: array
      items:
        $ref: "#/components/schemas/UserCardInfo"
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ar_Auth